feat: fix minor errors in tests(closes #35)
